% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdmeta.R
\name{plot.cdmeta}
\alias{plot.cdmeta}
\title{Plot the predictive distribution from a cdmeta object}
\usage{
\method{plot}{cdmeta}(
  x,
  show_interval = TRUE,
  breaks = 40,
  main = NULL,
  xlab = NULL,
  hist_col = "grey85",
  border_col = "white",
  density_lwd = 2,
  pi_col = "firebrick",
  pi_lwd = 3,
  rug = FALSE,
  transf = NULL,
  transf_name = NULL,
  qtype = 8,
  ...
)
}
\arguments{
\item{x}{An object of class \code{"cdmeta"}.}

\item{show_interval}{A logical value indicating whether a separate prediction-
interval panel is shown below the predictive distribution.}

\item{breaks}{Number of histogram breaks or another value accepted by
\code{graphics::hist()}.}

\item{main}{Optional main title.}

\item{xlab}{Optional x-axis label.}

\item{hist_col}{Fill color of histogram bars.}

\item{border_col}{Border color of histogram bars.}

\item{density_lwd}{Line width of the kernel density curve.}

\item{pi_col}{Color used for prediction-interval and predictive-estimate marks.}

\item{pi_lwd}{Line width used for the prediction interval.}

\item{rug}{A logical value indicating whether a rug of Monte Carlo draws is
added.}

\item{transf}{Optional transformation function for the predictive draws.}

\item{transf_name}{Optional character string naming \code{transf}.}

\item{qtype}{The quantile type used in \code{stats::quantile()}. Default is 8.}

\item{...}{Additional arguments passed to \code{graphics::hist()}.}
}
\value{
The input object \code{x}, invisibly.
}
\description{
Displays the Monte Carlo predictive distribution of the true effect in a
future study and, optionally, its prediction interval.
}
\examples{
data(hf_iron)
fit_hf <- cdmeta(hf_iron$yi, hf_iron$sei, B = 10000, seed = 11111)
plot(fit_hf)

}
